requirements.txt
numpy
pandas
scikit-learn
torch
python-telegram-bot
